+++ src/main/java/com/odont/odont/models/dao/IPersonDao.java
@@ -5,7 +5,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-public interface IPersonDao extends CrudRepository  < PersonEntity, Long> {
+public interface IPersonDao extends CrudRepository  < PersonEntity, Integer> {
 
     List<PersonEntity> findAllByStatus(int status);
 }
